requirements.txt
apache-libcloud==3.0.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
cloudpickle==1.4.1
future==0.18.2
futures==3.1.1
idna==2.9
numpy==1.18.4
opencv-python==4.2.0.34
Pillow==7.1.2
py-tes==0.4.0
PyMuPDF==1.17.0
pytesseract==0.3.4
python-dateutil==2.6.1
requests==2.23.0
six==1.15.0
urllib3==1.25.9
Wand==0.6.1
